upwork-devs/samarth_dd/storage_adapter/storage/requirements.txt
minio==6.0.0
python-dotenv == 0.13.0
nose          ==1.3.7
pytz==2020.1
Twisted==20.3.0